montblanc-dashboard/requirements.txt
flask>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
matplotlib>=3.7.0
Werkzeug>=2.3.0
numpy>=1.24.0
